fix: OpenCloseModal form story no longer keeps state after closing.
In the story the `<Form>` is always rendered this has the unintended
side-effect that the `<Form>` will keep track of the forms data even
when the modal is closed. This is strange behaviour, as the end user
expect the modal to start fresh each time.

The solution was to not render the `<Form>` when the `isOpen` `useState`
is not `true`. This way the `<Form>` is always reset when closing and
opening it.

Fixes #496