chore: updated the TypeaheadMultiple and TypeaheadSingle, changed Tag.
Updated `react-bootstrap-typeahead` from `4.0.0-alpha.6` to `5.1.2`.
Also updated the `@types/react-bootstrap-typeahead` from `3.4.5` to
`3.4.7`.

By upgrading ot the latest `react-bootstrap-typeahead` the bootstrap 4
style no longer needs to be imported in `main.scss` as it is now the
default style.

Modernized the `TypeaheadSingle` and `TypeaheadMultiple`:

1. Changed from a class component to a function component.

2. Modernized the tests: the `setup` function now returns and object
   with the shallow rendered component and all spies. Also added better
   tests for testing if the `loading` property is set correctly.

Altered the `Tag` component, before it had a strange relationship with
the `TypeaheadMultiple`. Originally the `Tag` component was created
to be used only inside of the `TypeaheadMultiple`, but in time it proved
to be very usable on its own. But because it was originally built for
the `TypeaheadMultiple` the SCSS of the `Tag` contained rules for the
`TypeaheadMultiple` and vice versa.

Now the `Tag` has been separated from the `TypeaheadMultiple`. This is
done by making the `TypeaheadMultiple` use the `renderToken` prop of the
`AsyncTypeahead` to render the `Tag` component. This way no convoluted SCSS
rules are needed. The `Tag` also does not need to render a `.token`
anymore for use in the `TypeaheadMultiple`, and needs not render an
extra `<span>` anymore.

Also cleaned up unused rules in the `Tag.scss` and `Typeahead.scss`,
and added comments to clarify rules that are still needed.