fix: fixed broken "npm run release" script
It did not commit the docs folder via `git add docs` for the storybook
changes. It was removed by me because I thought it was not needed.
Added it back but this time in the `storybook:build` task so it is
clear which component (storybook) causes these files to exist / change.

Also moved the `git clean -df` to the `storybook:build` task for the
same reason. The `git clean -df` is needed to remove `d.ts` definitions
which storybook generates for some reason.