chore: upgraded all packages.
Upgraded to the latest TypeScript: `4.1.2`. Changed all `interface` to
`type` where possible.

Due to upgrade in `final-form` the validators now have a new type, we
should no longer include `undefined` or `null` in the types.

Added `jest` and 'react to the `eslintrc`. Fixed issues coming out of
the new lint rules.

Added `jest-watch-typeahead` for easier searching when running jest in
interactive mode.

For storybook had to update our `show-code` fix to work with the new
version. Also updated the story separators from `|` to `/`.

Removed empty if statement in `layout.ts` in the `handleEpicDetailRow`
function.

The published package no longer includes test or story `d.ts` files.