improvement: removed DateRangePicker for helper functions.
Reasons for removing the `DateRangePicker` / `JarbDateRangePicker`.

1. The `DateRangePicker` is very inflexible UI wise it is not possible
   to have a start date, reminder date and an end date in that order.
   Because the start and end date are always next to each other either
   vertically or horizontally.

2. `JarbDateRangePicker` does not support setting both fields. It will
    only set the "start" field because only one jarb validator is
    allowed.

3. `JarbDateRangePicker` does not support range semantic errors as
   final form errors. This causes errors not to be picked up by
   final form, which is unexpected.

Instead I've added six new helper functions to create ranges using the
`DateTimeInput`.

The first three are for the `DateTimeInput` component `isDateAllowed`
prop:

1. `isDateBefore` whether or not the date is before the before date.
2. `isDateAfter` whether or not the date is after the after date.
3. `isDateBetween` whether or not the date is between the start and end.

The last three are for the `JarbDateTimeInput` component `validators`
prop. They are final form field validators. They use the first three
helpers:

4. `isDateBeforeValidator` uses `isDateBefore` to create a validator.
5. `isDateAfterValidator` uses `isDateAfter` to create a validator.
6. `isDateBetweenValidator uses `isDateBetween` to create a validator.

Updated the translations with new error messages for these validators.

It is possible to make the helpers inclusive or exclusive. This means
that for example in the `isDateAfter` you can either:

1. Make it accept only the dates after the date. This is `exclusive` and
   this the default.

2. Make it accept the date itself and all dates after the date. This is
   `inclusive` and must be configured as `exclusive` is the default.

Also the `DateTimeInput` now sends `undefined` instead of `null` when
the form element has no value. This way it is slightly nicer to use
in a `setState` which start as `undefined`.

Added example range stories, and updated the form example so it has
the `values` to be used for a range example by making it use a
final-form directly.

Closes: #457