improvement: the isDateBetween and isDateBetweenValidator work better
Before the `isDateBetween` and `isDateBetweenValidator` would only
check if both the start and end date were provided. Now they work when
either are provided.

The reason that this can work is that when the start is 2020-01-01 we
and there is no end. A date of 2000-01-01 can never pass because it lies
before the start date. The same is true for when you only have an end
date and no start date.