improvement: Button's now accept margin top and bottom.
Before the `Button` component would not accept `my`, `mt` or `mb`
bootstrap margin classes because a `<span>` is rendered which has
`display: inline;`

The reason for this is because `display: inline;` cannot use top or
bottom margins. Now it is a `display: inline-block` which does support
top and bottom margins when not rendered at `fullWidth`, and `d-block`
when rendered as `fullWidth`.

Now the `Button` is rendered as an `

Also fixed the full-width mode of the `Button` icon it misspelled:
`justify-content-start` as `justify-contents-start` and
`justify-end-start` as `justify-contents-end`.

Closes #458