improvement: now exposing more types and utils. Changed names of types.
The changing of the type names is a breaking change for people which
imported these types.

Before this commit the "inner" types of the components would be for
example: `IconPosition` for the `Button` component. This had the side
effect of names not being very "importable" because this would cause
collisions.

Now the inner types are all prefixed with the name of the component.
This means that `IconPosition` is now `ButtonIconPosition`.

Removed `UnionKeys` and `DistributiveOmit` types as they are not used
in the 42 ui itself, and were not exposed.

Also made the `validators` of `DateTimeInput` of type
`FieldValidator<Date>` instead of `FieldValidator<MomentInput>` as
that is the true type.