improvement: option to hide clear button
Sometimes a field is required and should work like a radio button, where
it is not possible to clear the field, you have to select at least one
option. It would be nice to be able to use all form components that way,
but for that we need an option whether to display the clear button.

Added canClear option to form components to be able to prevent users
from clearing the form element.

Closes #502