fix: EpicTable detail overlay padding
When using for instance a button as last element in the detail overlay
of the EpicTable, the button sticks to the bottom of the table, even
when a margin is used.

Added padding to the bottom of the detail overlay to prevent elements to
stick to the bottom of the detail overlay and allow for margins to add
extra spacing.

Closes #514