improvement: typeahead invert highlight styling
Users are aware of the text they just typed, so only emphasise the
predictive aspect of the query suggestion (and not what the users has
typed already, as is currently the case).

Inverted the font weight of the options to display the matching text in
normal font and the rest of the text in bold.

#517