improvement: ModalPickerMultiple without selection
The selection ribbon is only displayed when there is a selection in the
modal, but that causes the options to shift down when an option is
selected. It would be better if the selection ribbon could stay with a
text telling no options is selected.

Removed check that prevents selection ribbon from being displayed.
Added text to be displayed in the selection ribbon when no option is
selected.

Closes #509