feature: modal with suspense
Right now we have a OpenCloseModal, we've seen how ever that it is often
not enough. We want to create a Modal abstraction which can accept a
header / footer and body and which will suspend by default.

Added new Modal component with optional header and footer and a suspense
surrounding the body.
Changed OpenCloseModal to extend the new Modal component.
Deprecated isOpen property in all modal components. The component should
not be responsible for when it is displayed or any fallback, the
developer should only render the component when it should be rendered.

Closes #505