feature: info tooltip
You sometimes want to explain the use of a certain field for people that
use a form for the first time, but you don't want to display that
information all the time because it increases the size of the form and
therefore decreases its usability. A tooltip might be the perfect
solution in that case.

Added InfoTooltip compenent to display an info icon with a tooltip.

Closes #533