improvement: Typeahead should display all options
When the options provided is an array, the options should be displayed
right away instead of waiting for user input to match any option. Not
displaying the available options right away is only useful when a
request to an API is used, to decrease the load on the API.

Changed the TypeaheadSingle and TypeaheadMultiple components to use the
Typeahead component from react-bootstrap-typeahead when the options
provided is an array to be able to display all options when the user did
not type anything in the input (yet).

Closes #544

Co-authored-by: Pouja Nikray <poujanik@gmail.com>