fix: CheckboxMultipleSelect columns have static width
In the CheckboxMultipleSelect, the columns have a static width. When the
component is used in smaller layouts, the component overlays the
elements next to it due to the static width.

Added max-width 100% to prevent out-of-bounds display.
Removed chunking to prevent loads of whitespace when the number of
options is just over 10. The developer implementing this component is
now responsible for the grid.

Closes #552