improvement: DateTime: Better indicate un-selectable days, months, years
If dates cannot be selected, because of the condition specified in
`isDateAllowed`, this was only partly shown to the user: The cursor
became disabled and on hovering no ring was shown.

However, it is more clear if the text is also greyed out. And for the
month and year views (zoomed-out views), the disabled state was missing
altogether. This has been corrected and the disabled behaviour is now
the same across all three zoom levels (years, months, days).