feat (Typeahead, ModalPicker): Configurable page size
Added the ability to specify the number of suggestions for
the `async` variants of `TypeaheadSingle` and `TypeaheadMultiple`.

When the `options` passed to `TypeaheadSingle` and `TypeaheadMultiple`
are an array (`predefined`), all possible suggestions are always shown.

Added the ability to change the page size of `ModalPickerMultiple` and
`ModalPickerSingle`, for some applications it is desired to have more
than 10 items per page on screen.