improvement: readme
The link to Verdaccio in the readme does not link to the Verdaccio
website. According to the readme, we also want contributors to use yarn
for development, which we don't. The test commands are incorrect and we
don't explain anything about storybook in the readme. Also, there should
be a banner in the readme, but the file was never included in the
repository.

Corrected the link to Verdaccio.
Changed the explanation in setup to use npm instead of yarn.
Added explanation about the use of storybook and how to run a local
environment.
Replaced non-existing banner image with 42 BV logo.