feature: tabs
It would be nice to have tabs where the content and navigation item are
together so you can disable a tab or specify if a user is allowed to see
a tab in one place without having to put the code all over the place.
The content should also be suspended to improve performance.

Added Tabs component.

Closes #569