feature: debug
When I am developing, I often want to know what a variable is at a
certain moment. To be able to know that, I could use console.log, but
then I have to search within all the console logging, or I could JSON
stringify the object. I would like to have a component to just pass the
variable I want to debug and it'll JSON stringify to make debugging
easier.

Added Debug component.

Closes #572