improvement: SubmitButton now scrolls to closest error.
The `SubmitButton` has a new prop now called `scrollToClosestError`
when `true`, which it is by default, whenever the button is clicked
and there are `FormError`'s the user is scrolled to the first error.

Also gave `SubmitButton` a new prop called `icon` so the icon can
optionally be configured. Done so the `OpenCloseModal` can now also use
the `SubmitButton`, and get the scrolling behavior as well.

The `Modal`'s story now uses the `SubmitButton` as well.

To get the `Modal`, and `OpenCloseModal` stories to scroll to the
error, the `FinalForm` helper util now no longer renders the
final-form `Form`. This way the stories will not close the modal on
submit automatically, but instead can be controlled to wait until
the form is valid.

Refactoring: removed `Form` from `story-utils.tsx` as it was a `Card`,
and the name was a bit misleading, replaced with simple `Card` call.
In the `FinalForm` util component I've removed the `Fragment` as it
as not needed, also it now uses the `Debug` component.

Closes #585