improvement: replace React Async with React Query
There is a dependency to React Async, which is used in the AsyncContent,
AsyncList and AsyncPage components. That library has not been touched in
a year and is prone to security issues. There are a few other libraries
like React Query or SWR which provide the same functionality and more.

Added support for React Query in Async components.

Closes #601