feature: error boundary in card
It would be nice if the card has an error boundary to allow rendering of
other parts of the application when an error occurs within the card.

Added ErrorBoundary component to allow developers to catch and display
errors without breaking the entire application.
Added ErrorBoundary to Card body to automatically catch and display
errors in cards.

Closes #596