improvement: prevent click event propagation
When an icon component has an onChange defined, the click event is not
passed, so preventDefault can not be called by the implementation.
Due to that, click events in parent/higher order components are
triggered as well.

Added preventDefault call on click events for the icon components to
prevent event propagation.

Closes #606