feature: EpicTable resize trigger
We have a case where the space the table has is not enough to fit all
columns, but when the sidebar is minified, there is enough space to fit
all columns. When the sidebar is opened again, the table is not resized
until the user moves the mouse over the EpicTable. We would like to have
a custom trigger we can call to force the table to resize.

Added event listener to listen to custom resizeEpicTables event as well
as window resize.

Closes #608