feature: CrudTable
We are using the EpicTable a lot for our entity listings and it would be
a lot easier if there is a component that simplifies the EpicTable for
that purpose.

Added CrudTable component which wraps the EpicTable to make it easier to
display a table with create/edit/detail view and pagination.
Added CrudHeader component to simplify the display of a header with
sort and search functionality.

Closes #605