improvement: Pagination static number of items
Currently, the pagination component could change position when it is
centered on a page depending on what page you are on. This should be
changed to always display a static amount of buttons, so the position of
the pagination component does not change when navigating between pages.

Changed pagination to always display the same amount of buttons.

Closes #604