improvement: select without placeholder
The placeholder you define cannot be disabled. If you don't specify a
placeholder, you still have an empty value you can select even if you
don't want to. It would help if Select would not display an empty option
if the placeholder is empty.

Added check if placeholder is empty to prevent an empty option.

Closes #616