improvement: accessibility
We would like our components to be accessible, so they could be used in
websites that have to comply with the WCAG standards.

Changed some (text) colors to improve contrast.
Changed label property of form elements to be required.
Added property to form elements to use invisible label (aria-label).
Added a11y addon to Storybook to be able to test and display WCAG 2
compatibility.