improvement: optional dependencies
We have a lot of dependencies that are only used by certain components,
but are defined as strong dependencies. We should move those
dependencies to optionalDependencies to minimize the dependency tree and
give the developer more control over what dependencies are used in their
project.

Moved dependencies that are only used by few components to
optionalDependencies.