improvement: ImageUpload user experience
Uploaded images should no longer be stretched out if they are smaller
than the configured format. For these, the cropping step should be
skipped.
When users selected an image, they often don't understand they have to
click the done button as well to save the image. If they like what they
see, they expect the image to be used as is automatically.

Added a check if the uploaded image is too small to skip the cropping
step and use the image as is.
Uploaded images are now saved as soon as they are put in, rather than
when a user is done with editing.

Closes #473
Closes #627