improvement: optional peer dependencies
We previously used peerDependencies, which caused the projects using
this library to require the listed peer dependencies, while they might
not actually be used. We switched to optionalDependencies, but
optionalDependencies are automatically installed, so the developer is
not responsible for installing those packages in that case either. We
need a way to actually make the developer responsible for installing the
dependencies only used by specific components.

Changed optionalDependencies back to peerDependencies and added
peerDependenciesMeta to specify which peer dependencies are actually
optional.