improvement: form in OpenCloseModal
We use the OpenCloseModal for forms, as a submit button is displayed in
the footer. There is no possibility however to wrap the modal with a
form tag, which causes invalid HTML or could cause unexpected behaviour
as the submit button could submit another wrapping form as well.

Changed OpenCloseModal to use custom modal with form in content wrapping
body and footer.

Closes #689