fix: pagination classes on wrapper
When custom classes are added to the Pagination element to position the
element, the positioning is done on the pagination items, but not on the
total elements or page size elements.

Changed Pagination components to add class to wrapper instead of
pagination element.
Added translation to customize page size dropdown options text.