fix: OpenCloseModal onSave type
The onSave function on OpenCloseModal is passed to onSubmit of the form.
The onSave does not expect any value, but the onSubmit sends a form
event. Without the event, we cannot prevent the HTML form from
automatically sending a request, reloading the page.

Changed onSave type to FormEventHandler to allow event handling.