fix: modal picker selection layout
The ModalPickerMultiple displays the current selection in an area with a
gray background that doesn't align with the borders of the modal.

Changed the modal padding to match the spacing of rows and columns.
Changed spacing of the gray row to match the alignment of the rest of
the modal content.
Changed gray row background to use bootstrap classes instead of custom
color.