improvement: reload options after value changes
When using an array of options for a choice form component, the
displayed list of options includes the selected value when it is not
present in the options array to still be able to see the selected value.
When the value is updated, the old value is still visible and the user
is able to choose that value, while it is not a valid option anymore.

Added listener to change the list of options when the value changes.