improvement: clearing DateTimeInput
Manually clearing the input doesn't work. Only of you clear the input
and change the date and clear it a second time, the value is actually
cleared.

Added clear button to clear the input as manually clearing the input is
an issue within the react-datetime library that is not updated anymore.
This allows for the DateTimeInput to support undefined in onChange.

Closes #704