improvement: column minimum width
When resizing a column, the minimum width causes the columns not to
become smaller than the specified width. The minimum width however is
based on the initial width of the column. Developers should be able to
specify the width and minWidth separately to be able to default to a
comfortable width while allowing to decrease the width.

Added minWidth property to allow specifying minimum width separately
from initial width.