improvement: display submit errors
Sometimes a form is submitted and the back-end does not return a
field-related error, but we want to display a custom error underneath a
specific field instead of having to display the error in another way.

Changed field error handling to also support submit errors returned by
the function specified in onSubmit.