improvement: epic selection spacing
When using EpicSelection in an EpicTable, there is just enough space
between the edges of the row and the EpicSelection. If the user clicks
that space, the row click is triggered, while the user wanted to select
the checkbox instead.

Added padding and negative margin to EpicSelection to negate cell
padding.