improvement: ModalPicker only fetch options when opened
When using a ModalPicker variant, the options are fetched before the
button is clicked to open the modal. This means an unnecessary request
might be performed to fetch options that are never used.

Moved options to from ModalPicker variants to ModalPicker itself and
conditionally render the ModalPicker to prevent pre-fetching options.
Added key and label to renderOptions callback.