fix: Epic table layout warped when using fragments
When using fragments in the EpicTable, for instance in the example with
the multiple header rows, the layout is warped. The rows become columns.

Changed react-is dependency to version compatible with the react version
being installed.
Changed react-is dependency to become optional and only to be required
when EpicTable is used.